Agregada la branch para realizar la 1era parte del trabajo Cambio de floats a ints ante error de user trap al ejecutar cpubound. Esto fue charlado por zulip con los profesores y deberá ser agregado como aclaración en el informe final

+++ user/cpubench.c
@@ -8,9 +8,9 @@
 #define TIMES 32
 #define MINTICKS 100
 
-static float a[N][N];
-static float b[N][N];
-static float c[N][N];
+static int a[N][N];
+static int b[N][N];
+static int c[N][N];
 
 static void init(void) {
   int x, y;
@@ -18,12 +18,12 @@ static void init(void) {
     for (x = 0; x < N; ++x) {
       a[y][x] = y - x;
       b[y][x] = x - y;
-      c[y][x] = 0.0f;
+      c[y][x] = 0;
     }
   }
 }
 
-static void matmul(float beta) {
+static void matmul(int beta) {
   int x, y, k;
   for (y = 0; y < N; ++y) {
     for (x = 0; x < N; ++x) {
@@ -38,7 +38,7 @@ int
 main(int argc, char *argv[])
 {
   int pid = getpid();
-  float beta = 1.0f;
+  int beta = 1;
 
   init();
   int start = uptime();
